Create and delete Commands completed

+++ Structs/Ebr.go
@@ -3,8 +3,8 @@ package Structs
 type Ebr struct {
 	Part_status byte
 	Part_fit    byte
-	Part_start  int
-	Part_size   int
+	Part_start  int64
+	Part_size   int64
 	Part_name   [16]byte
-	part_next   int
+	part_next   int64
 }
