+++ Structs/Partition.go
@@ -4,7 +4,7 @@ type Partition struct {
 	Part_status byte
 	Part_type   byte
 	Part_fit    byte
-	Part_start  int
-	Part_size   int
+	Part_start  int64
+	Part_size   int64
 	Part_name   [16]byte
 }
